fix deepResolve to ignore streamable, add openai vnext playground
